Finishing up adaptor testing
- Got a few more tests working
- Moved miniio tests to a new file for minio adaptor testing
- Left a TOOD - we need to get auth process tested
